+++ Source/MTAFRAT/plugins/3D Torrents/3D Torrents.vb
@@ -16,9 +16,9 @@ Class _3DTorrentsPlugin : Inherits DynamicPlugin
     Overloads Async Function RunAsync() As Task(Of RegistrationStatus)
         Return Await Task.Run(
             Function()
-                Using driver As ChromeDriver =
-                    CreateChromeDriver(Me, headless:=True,
-                                       $"--unsafely-treat-insecure-origin-as-secure={Me.Url}") ' Required for insecure "http" protocol.
+                Using service As ChromeDriverService = Nothing,
+                      driver As ChromeDriver = CreateChromeDriver(Me, service, headless:=True,
+                                                                  $"--unsafely-treat-insecure-origin-as-secure={Me.Url}")
                     Try
                         LogMessageFormat(Me, "StatusMsg_ConnectingFormat", Me.Name)
                         NavigateTo(driver, Me.Url)