+++ Source/MTAFRAT/plugins/AvistaZ/AvistaZ.vb
@@ -16,11 +16,12 @@ Class AvistaZPlugin : Inherits DynamicPlugin
     Overloads Async Function RunAsync() As Task(Of RegistrationStatus)
         Return Await Task.Run(
             Function()
-                Using driver As ChromeDriver = CreateChromeDriver(Me, headless:=False)
+                Using service As ChromeDriverService = Nothing,
+                      driver As ChromeDriver = CreateChromeDriver(Me, service, headless:=False)
                     Try
                         LogMessageFormat(Me, "StatusMsg_ConnectingFormat", Me.Name)
                         NavigateTo(driver, Me.Url)
-                        
+
                         LogMessage(Me, "StatusMsg_CloudflareTrialWait")
                         WaitForPageReady(driver, TimeSpan.FromSeconds(10))
                         Thread.Sleep(5000)